feat: repackage UiText to model module
