+++ core/ui/src/main/kotlin/com/divinelink/core/ui/components/AppTopAppBar.kt
@@ -21,9 +21,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
+import com.divinelink.core.model.UIText
 import com.divinelink.core.ui.R
 import com.divinelink.core.ui.TestTags
-import com.divinelink.core.ui.UIText
 import com.divinelink.core.ui.getString
 
 @OptIn(ExperimentalMaterial3Api::class)
