+++ core/ui/src/main/kotlin/com/divinelink/core/ui/components/dialog/AlertDialogUiState.kt
@@ -1,7 +1,7 @@
 package com.divinelink.core.ui.components.dialog
 
+import com.divinelink.core.model.UIText
 import com.divinelink.core.ui.R
-import com.divinelink.core.ui.UIText
 
 data class AlertDialogUiState(
   val title: UIText = UIText.ResourceText(R.string.general_error_title),
