+++ core/ui/src/main/kotlin/com/divinelink/core/ui/components/expandablefab/ExpandableFloatingActionButton.kt
@@ -47,10 +47,10 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.IntOffset
 import com.divinelink.core.designsystem.theme.AppTheme
 import com.divinelink.core.designsystem.theme.dimensions
+import com.divinelink.core.model.UIText
 import com.divinelink.core.ui.IconWrapper
 import com.divinelink.core.ui.Previews
 import com.divinelink.core.ui.TestTags
-import com.divinelink.core.ui.UIText
 import com.divinelink.core.ui.getString
 import com.divinelink.core.ui.snackbar.controller.LocalSnackbarController
 import com.divinelink.core.ui.snackbar.controller.ProvideSnackbarController
