+++ core/ui/src/main/kotlin/com/divinelink/core/ui/components/expandablefab/FloatingActionButtonItem.kt
@@ -1,7 +1,7 @@
 package com.divinelink.core.ui.components.expandablefab
 
+import com.divinelink.core.model.UIText
 import com.divinelink.core.ui.IconWrapper
-import com.divinelink.core.ui.UIText
 
 open class FloatingActionButtonItem(
   val icon: IconWrapper,
