+++ core/ui/src/main/kotlin/com/divinelink/core/ui/snackbar/SnackbarMessage.kt
@@ -3,7 +3,7 @@ package com.divinelink.core.ui.snackbar
 import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarResult
 import androidx.compose.material3.SnackbarVisuals
-import com.divinelink.core.ui.UIText
+import com.divinelink.core.model.UIText
 
 sealed interface SnackbarMessage {
