+++ feature/details/src/main/kotlin/com/divinelink/feature/details/person/ui/PersonalInfoSectionData.kt
@@ -1,6 +1,6 @@
 package com.divinelink.feature.details.person.ui
 
-import com.divinelink.core.ui.UIText
+import com.divinelink.core.model.UIText
 
 data class PersonalInfoSectionData(
   val title: UIText,
