+++ feature/onboarding/src/main/res/values/strings.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-  <string name="feature_onboarding_next">Next</string>
   <string name="feature_onboarding_get_started">Get started</string>
   <string name="feature_onboarding_skip">Skip</string>
 
@@ -9,10 +8,8 @@
 
   <string name="feature_onboarding_tmdb_page_title">Sync with TMDB</string>
   <string name="feature_onboarding_tmdb_page_description">Connect your TMDB account to access your watchlists and ratings.</string>
-  <string name="feature_onboarding_tmdb_page_action">Connect TMDB Account</string>
-  <string name="feature_onboarding_successfully_connected">Connected successfully</string>
+  <string name="core_model_onboarding_tmdb_page_action">Connect TMDB Account</string>
 
   <string name="feature_onboarding_jellyseerr_page_title">Jellyseerr Integration</string>
   <string name="feature_onboarding_jellyseerr_page_description">If you have access to a Jellyseerr server, connect your account to request movies and TV shows instantly.</string>
-  <string name="feature_onboarding_jellyseerr_page_action">Link Jellyseerr Account</string>
 </resources>
